Added `ipfs refs`
I got a malformed syntax error when trying refs on Postman, but it seemed to work fine in the command line. I am not sure what exactly is happening with the format option, and I suspect that coud be improved a bit. Otherwise, this seems fine to me. Threw in headers because I might as well, they were right there.