Added Group `pin`
I am not sure about ls count. I do not think that ls --quiet is an argument: the go code is a flag for Fprintf or not, which makes me think it does not affect the API returns. But I cannot tell, because in Wireshark it is all sent as flags.